Reload page on cube save to update tree.
Fixed graph generation to to use parseFloat

+++ web/js/expInterface.js
@@ -87,7 +87,7 @@ $("document").ready(function(){
                                                 var y = [];
                                                 for(var i = 0; i < results[0].length; i++)
                                                 {
-                                                    y.push(parseInt(results[1][i]));
+                                                    y.push(parseFloat(results[1][i]));
                                                 }
                                                 var chart = new Highcharts.Chart({
                                                     chart: {
